% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeDifferentialExpression.R
\name{analyzeDifferentialExpression}
\alias{analyzeDifferentialExpression}
\title{Analyze Differential Gene Expression}
\usage{
analyzeDifferentialExpression(
  expressionDataPath,
  significantDEGsOutputPath,
  allDEGsOutputPath,
  groupClassification
)
}
\arguments{
\item{expressionDataPath}{Path to the expression counts file.}

\item{significantDEGsOutputPath}{Path to the .xlsx file that will be created,
containing a list of most significant DEGs.}

\item{allDEGsOutputPath}{Path to the .xlsx file used as a universe in the
other functions in the program, containing a list of all DEGs unfiltered.}

\item{groupClassification}{A vector of ones and zeros specifying which samples
(columns) are control/healthy (0) and which are disease (1).}
}
\description{
This function filters and normalizes RNA-seq expression data, calculates
statistically significant differentially expressed genes (DEGs), and exports
them to an Excel file. It filters DEGs based on FDR < 0.05 and absolute logFC value > 1.3.
The low expression filtering parameters are set to keep genes with at least 10 counts
in at least 3 samples.
}
\examples{
analyzeDifferentialExpression("E-MTAB-2523.counts.txt",
                              "Significant_DEGs_E-MTAB-2523.xlsx",
                              "Complete_DEG_List.xlsx",
                              c(1,0,1,1,0,1,0,1,0,1,1,1,0,1,1,1,0,1))
}
