% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichGONE.R
\name{enrichGONE}
\alias{enrichGONE}
\title{Perform Gene Ontology Enrichment Analysis}
\usage{
enrichGONE(Input_File_1, Input_File_2, Output_File)
}
\arguments{
\item{Input_File_1}{Path to the .xlsx file containing differentially expressed genes (DEGs) for GO enrichment analysis.}

\item{Input_File_2}{Path to the .xlsx file containing universe genes for the analysis.}

\item{Output_File}{Path to save the results of the enrichGO analysis in .xlsx format.}
}
\value{
This function does not return a value but generates plots and writes the results of the enrichment analysis to an Excel file.
}
\description{
This function performs gene ontology (GO) enrichment analysis using the clusterProfiler package. It reads gene expression data from specified input files, converts gene symbols to ENTREZ IDs, performs enrichment analysis, generates various plots, and exports the results to an Excel file.
}
\examples{
enrichGONE("path/to/DEGs.xlsx", "path/to/universe_genes.xlsx", "path/to/enrichGONE_results.xlsx")
}
