% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performGeneEnrichmentAnalysis.R
\name{performGeneEnrichmentAnalysis}
\alias{performGeneEnrichmentAnalysis}
\title{Gene Enrichment Analysis}
\usage{
performGeneEnrichmentAnalysis(
  genesFilePath,
  universeFilePath,
  enrichmentResultsPath
)
}
\arguments{
\item{genesFilePath}{Path to the DEGs file for genes input in enrichGO.}

\item{universeFilePath}{Path to the DEGs file for universe input in enrichGO.}

\item{enrichmentResultsPath}{Path to the output .xlsx file containing enrichGO analysis results.}
}
\description{
Prepares data from DEG analysis for Gene Enrichment Analysis using enrichGO from clusterprofiler.
It uses two files: one for DEGs (used in the 'genes' attribute of enrichGO) and one for the universe parameter, both including log fold change, p-adjustment, and FDR correction.
Generates barplot, dotplot, and upset plot for enrichment visualization.
}
\examples{
library(DEGpathwayAnalysis)
performGeneEnrichmentAnalysis("processed_data/DEGs_from_E-MTAB-2523.xlsx",
                              "processed_data/universe.xlsx",
                              "processed_data/gene_enrichment_results.xlsx")

}
