% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performKEGGAnalysis.R
\name{performKEGGAnalysis}
\alias{performKEGGAnalysis}
\title{KEGG Pathway Enrichment Analysis}
\usage{
performKEGGAnalysis(genesDataPath, universeDataPath, keggResultsPath)
}
\arguments{
\item{genesDataPath}{Path to the DEGs file for gene input in enrichKEGG.}

\item{universeDataPath}{Path to the DEGs file for universe input in enrichKEGG.}

\item{keggResultsPath}{Path to the .xlsx file that will be created, containing enrichKEGG analysis results.}
}
\description{
Conducts KEGG Pathway Enrichment Analysis on DEGs using enrichKEGG from the clusterProfiler package.
Utilizes filtered and log fold change adjusted DEG data files for gene input and universe input in enrichKEGG.
Generates barplot, dotplot, and upset plot for enrichment visualization.
}
\examples{
library(DEGpathwayAnalysis)
performKEGGAnalysis("processed_data/DEGs_from_E-MTAB-2523.xlsx",
                    "processed_data/universe.xlsx",
                    "processed_data/kegg_analysis_results.xlsx")

}
